+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/utils/AppInfoUtil.kt
@@ -6,25 +6,51 @@
 package com.amazonaws.services.chime.sdk.meetings.internal.utils
 
 import android.content.Context
+import com.xodee.client.audio.audioclient.AppInfo
 import com.xodee.client.video.VideoClient
 
 object AppInfoUtil {
-    fun initializeVideoClientAppDetailedInfo(context: Context) {
-        val manufacturer = DeviceUtils.deviceManufacturer
-        val model = DeviceUtils.deviceModel
-        val osVersion = DeviceUtils.osVersion
+    private lateinit var manufacturer: String
+    private lateinit var model: String
+    private lateinit var osVersion: String
+    private lateinit var appName: String
+    private lateinit var appCode: String
+    private const val clientSource = "amazon-chime-sdk"
+    private lateinit var sdkVersion: String
+
+    private fun initializeAppInfo(context: Context) {
+        manufacturer = DeviceUtils.deviceManufacturer
+        model = DeviceUtils.deviceModel
+        osVersion = DeviceUtils.osVersion
         val packageName = context.packageName
         val packageInfo = context.packageManager.getPackageInfo(packageName, 0)
-        val appVer = packageInfo.versionName
-        val appCode = packageInfo.versionCode.toString()
-        val clientSource = "amazon-chime-sdk"
-        val sdkVersion = DeviceUtils.sdkVersion
+        appName = String.format("Android %s", packageInfo.versionName)
+        appCode = packageInfo.versionCode.toString()
+        sdkVersion = DeviceUtils.sdkVersion
+    }
+
+    fun initializeVideoClientAppDetailedInfo(context: Context) {
+        initializeAppInfo(context)
 
         VideoClient.AppDetailedInfo.initialize(
-            String.format("Android %s", appVer),
+            appName,
             appCode,
             model,
             manufacturer,
+            osVersion,
+            clientSource,
+            sdkVersion
+        )
+    }
+
+    fun initializeAudioClientAppInfo(context: Context): AppInfo {
+        initializeAppInfo(context)
+
+        return AppInfo(
+            appName,
+            appCode,
+            manufacturer,
+            model,
             osVersion,
             clientSource,
             sdkVersion